// front-end/src/pages/PagesForMobile/PersonalAccountMobile/components/Middle/types.ts
export interface IChangedInfoData {
  id: string
  name: string
  lastName: string
  email: string
  telefone: string
}
export interface IChangedPassword {
  oldPassword: string
  newPassword: string
  repeatNewPassword: string
}

export interface IAllDeliveryAddress {
  id: string
  numberInAddress: string
  deliveryAddress: string
}